refactor(client-v2): share route runtime version detection
