+++ packages/core/client-v2/src/__tests__/getRouteRuntimeVersion.test.ts
@@ -1,0 +1,68 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+import { afterEach, describe, expect, it } from 'vitest';
+import { getRouteRuntimeVersion } from '../utils/getRouteRuntimeVersion';
+
+function replacePathname(pathname: string) {
+  Object.defineProperty(window, 'location', {
+    configurable: true,
+    value: {
+      ...window.location,
+      pathname,
+    },
+  });
+}
+
+describe('getRouteRuntimeVersion', () => {
+  afterEach(() => {
+    delete window.__nocobase_modern_client_prefix__;
+    delete window.__nocobase_public_path__;
+    replacePathname('/');
+  });
+
+  it('returns modern when the current pathname is inside the modern client prefix', () => {
+    window.__nocobase_modern_client_prefix__ = 'v';
+    window.__nocobase_public_path__ = '/nocobase/v/';
+    replacePathname('/nocobase/v/admin/settings/workflow');
+
+    expect(getRouteRuntimeVersion()).toBe('modern');
+  });
+
+  it('returns legacy when reused v2 components run under the v1 settings page', () => {
+    window.__nocobase_modern_client_prefix__ = 'v';
+    window.__nocobase_public_path__ = '/nocobase/';
+    replacePathname('/nocobase/admin/settings/workflow/workflows/123');
+
+    expect(getRouteRuntimeVersion()).toBe('legacy');
+  });
+
+  it('returns legacy when root public path is "/"', () => {
+    window.__nocobase_modern_client_prefix__ = 'v';
+    window.__nocobase_public_path__ = '/';
+    replacePathname('/admin/settings/workflow/workflows/123');
+
+    expect(getRouteRuntimeVersion()).toBe('legacy');
+  });
+
+  it('returns modern when root public path is "/v/"', () => {
+    window.__nocobase_modern_client_prefix__ = 'v';
+    window.__nocobase_public_path__ = '/v/';
+    replacePathname('/v/admin/workflow/workflows/123');
+
+    expect(getRouteRuntimeVersion()).toBe('modern');
+  });
+
+  it('falls back to pathname matching when public path is unavailable', () => {
+    window.__nocobase_modern_client_prefix__ = 'v';
+    replacePathname('/v/admin/workflow/workflows/123');
+
+    expect(getRouteRuntimeVersion()).toBe('modern');
+  });
+});
